requirements.txt
flask==2.0.3
flask-cors==3.0.10
firebase-admin==5.2.0
anthropic==0.30.1
python-dotenv==0.19.2
gunicorn==20.1.0
Werkzeug==2.0.3